requirements.txt
appnope==0.1.0
astroid==2.4.2
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.2
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pyzmq==19.0.2
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
